Separated Twitch API calls from polybot.js completely. Still testing.
