+++ polybot.js
@@ -38,33 +38,10 @@ let commandFunctions = commands.map((command) => {
     return require(`./commands/${command}.js`);
 });
 
-
-// New Twitch API
-const querystring = require("querystring"),
-  fetch = require("node-fetch");
-
-const streams_url = "https://api.twitch.tv/helix/streams";
-
-// JSON objects - These are the custom query strings for the streams request
-const qs = querystring.stringify({
-      user_id: "92433498"
-});
-
-// Template Strings used to make the streams URL w/ our custom query strings
-const qURL = `${streams_url}?${qs}`;
-
-// Adds the client id as a header for the fetch
-const fetchArgs = {
-    headers: {
-        "Client-ID": client_id
-    }
-};
-
-// Final fetch API call
-fetch(qURL, fetchArgs)
-    .then(response => response.json()) // Get JSON Response
-    .then(data => startTime = new Date(data.data[0].started_at)) // View the Retrieved Data
-    .catch(error => console.error(error)); // Catch any Errors
+// Initialize getStartTime function
+const getStartTime = require('./helpers/getStartTime.js');
+// Attempt to get startTime when bot starts
+getStartTime();
 
 
 // ChatBot Configuration Options
@@ -120,9 +97,13 @@ client.on('chat', (channel, user, message, self) => {
     commandFunctions.forEach((command) => {
       if (typeof command === "function") {
         if (command.name === "uptimeCommand") {
-          if (typeof startTime === "undefined") {
-            return "startTime is not defined yet";
-          }
+            if (typeof startTime === "undefined") {
+                let response = command(trimmedMessage);
+                if(response){
+                client.action('epistemicpolymath', response);
+                }
+                return;
+            }
             let response = command(trimmedMessage, startTime);
             if(response) {
               client.action('epistemicpolymath', response);
@@ -130,7 +111,7 @@ client.on('chat', (channel, user, message, self) => {
               return;
         }
         let response = command(trimmedMessage, user, channel);
-          if(response){
+          if(response) {
               client.action('epistemicpolymath', response);
           }
       }
